requirements.txt
asgiref==3.8.1
astroid==3.3.6
dill==0.3.9
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
PyJWT==2.10.1
pylint==3.3.2
sqlparse==0.5.2
tomlkit==0.13.2
